\name{disease_class}
\alias{disease_class}
\docType{data}
\title{
same as lls
}
\description{
same as lls
}
\usage{data("disease_class")}
\format{
  A data frame with 39 observations on the following 2 variables.
  \describe{
    \item{\code{sample}}{a character vector}
    \item{\code{Condations}}{a character vector}
  }
}
\details{
same as lls
}
\source{
same as lls
}
\references{
same as lls
}
\examples{
data(disease_class)
## maybe str(disease_class) ; plot(disease_class) ...
}
\keyword{datasets}
