\name{metatopics-package}
\alias{metatopics-package}
\alias{metatopics}
\docType{package}
\title{
This is package for topic model application on microbial
}
\description{
This package utilizes unsupervised topic model method to understand the meta data of microbial. Additional visualization methods could act as an assistant for explore the relationship between returned topics and disease types and the relationship among microbial.
}
\details{
\tabular{ll}{
Package: \tab metatopics\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-13\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Todd Tao Qi <qitao052@126.com>
}
\references{

}

\keyword{ package }
\seealso{
see also package BiotypeR in github and package topicmodels in CRAN
}
\examples{
}
