\name{plot_perplexity}
\alias{plot_perplexity}

\title{
Plot the perplexity and likelihood returned from cross validation
}
\description{
This function plot the perplexity and likelihood returned from the function selectK, which is a function applying cross validation on topic models, in order to select a appropriate topic number.
}
\usage{
plot_perplexity(model, k)
}

\arguments{
  \item{model}{
A list returned from function \code{selectk}
}
  \item{k}{
A vector with integer as the number of topics. See parameter \code{kv} in function \code{selectK}
}
}
\details{

}
\value{

}
\references{

}
\author{

}
\note{

}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
#See example in function selectK}
