\name{topic_i_term}
\alias{topic_i_term}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
function used in function plot_network
}
\description{
function used in function plot_network
}
\usage{
topic_i_term(i, term_matrix, prob = 0.01)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{i}{
no avaliable
}
  \item{term_matrix}{
no avaliable
}
  \item{prob}{
no avaliable
}
}
\details{
no avaliable
}
\value{
no avaliable
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}